Convert/Update all remaining modules to be deterministic. Rebases previous:
Converted gender modules to be determinstic.

Convert remaining ipsum modules to be be deterministic.

Convert SSN modules to be deterministic.

Convert units and time modules to be deterministic.

Convert core string module to be deterministic.

Updated job* modules to be deterministic.

Updated currency and product to be deterministic, added test files.

Updated remaining "C*" modules to be determinisic.

Updated remaining "A*" modules to be deterministic.

Updated remaining "B*" modules to be deterministic.

Updated all name modules to be deterministic.

Update all identification modules to be deterministic.

Update all phone number modules to be deterministic.

Updated all remaining modules to be deterministic.