Update the Rakefile to not generaet a gemspec
It no longer uses jeweler, so shouldn't need to be regenerating the file.
Best practice is to have `git ls-files` in the gemspec itself. It will
be evaluated at the time that the gem is built, and an expanded version
of the gemspec is included in the gem.

This changes the Rakefile to remove the gemspec task, and move the logic
for determining files into the gemspec.

Some other changes related:

- remove check for VERSION file, since no longer using jeweler
- use `gem build` with `--output` to not require mving it after it's built
- add `-Ilib` to `reference` task, otherwise it fails to find ffaker