Deterministically generate REFERENCE.md
Right now, if you generate the file multiple times, you will different
results. This is noteworthy because it is generated as part of
`rake build` and `rake release`.

This is fixed by using assigning a `FFaker::Random.seed`.

Not specifically about a consistent REFERENCE.md, but I've also added
`unique` to the calls to prevent duplicates to get more variety.